+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - allow empty account name on msg.Validate()
 - enable fees for all domain module handlers
+- remove certificate indexing
+- fix account transfer
 
 ## 0.2.3